octets: make kib/mib generic via sealed Octets trait

+++ rama-cli/src/cmd/serve/httptest/endpoint/octet_stream.rs
@@ -17,7 +17,7 @@ pub(in crate::cmd::serve::httptest) fn service()
 -> impl Service<Request, Output = Response, Error = Infallible> {
     (
         ConsumeErrLayer::trace_as_debug(),
-        BodyLimitLayer::new(mib(8) as usize),
+        BodyLimitLayer::new(mib(8_usize)),
         MapResponseBodyLayer::new_boxed_streaming_body(),
     )
         .into_layer(service_fn(async |req: Request| {
